requirement.txt
Flask
slack_sdk
google-cloud-aiplatform